+++ pyogrio/_io.pyx
@@ -10,12 +10,14 @@ import locale
 import logging
 import math
 import os
+import sys
 import warnings
 
 from libc.stdint cimport uint8_t, uintptr_t
 from libc.stdlib cimport malloc, free
 from libc.string cimport strlen
 from libc.math cimport isnan
+from cpython.pycapsule cimport PyCapsule_GetPointer
 
 cimport cython
 import numpy as np
@@ -2178,3 +2180,182 @@ def ogr_write(
         exc = exc_check()
         if exc:
             raise DataSourceError(f"Failed to write features to dataset {path}; {exc}")
+
+
+def ogr_write_arrow(
+    str path,
+    str layer,
+    str driver,
+    arrow_obj,
+    str crs,
+    str geometry_type,
+    str geometry_name,
+    str encoding,
+    object dataset_kwargs,
+    object layer_kwargs,
+    bint append=False,
+    dataset_metadata=None,
+    layer_metadata=None,
+):
+    IF CTE_GDAL_VERSION < (3, 8, 0):
+        raise RuntimeError("Need GDAL>=3.8 for Arrow write support")
+
+    cdef OGRDataSourceH ogr_dataset = NULL
+    cdef OGRLayerH ogr_layer = NULL
+
+    if not encoding:
+        encoding = locale.getpreferredencoding()
+
+    layer_created = create_ogr_dataset_layer(
+        path, layer, driver, crs, geometry_type, encoding,
+        dataset_kwargs, layer_kwargs, append,
+        dataset_metadata, layer_metadata,
+        &ogr_dataset, &ogr_layer,
+    )
+    if geometry_name:
+        opts = {"GEOMETRY_NAME": geometry_name}
+    else:
+        opts = {}
+    cdef char **options = dict_to_options(opts)
+
+    stream_capsule = arrow_obj.__arrow_c_stream__()
+    write_arrow_stream_capsule(ogr_layer, stream_capsule, geometry_name, options)
+
+    ### Final cleanup
+    if ogr_dataset != NULL:
+        GDALClose(ogr_dataset)
+
+        # GDAL will set an error if there was an error writing the data source
+        # on close
+        exc = exc_check()
+        if exc:
+            raise DataSourceError(f"Failed to write features to dataset {path}; {exc}")
+
+
+IF CTE_GDAL_VERSION >= (3, 8, 0):
+
+    cdef write_arrow_stream_capsule(OGRLayerH destLayer, object capsule, str geometry_name, char** options = NULL):
+        cdef ArrowSchema schema
+        cdef ArrowArray array
+
+        cdef ArrowArrayStream* stream = <ArrowArrayStream*>PyCapsule_GetPointer(
+            capsule, "arrow_array_stream"
+        )
+        if stream == NULL:
+            raise RuntimeError("No valid stream.")
+
+        if stream.release == NULL:
+            raise RuntimeError("Arrow Array Stream was already released.")
+
+        errcode = stream.get_schema(stream, &schema)
+        if errcode != 0:
+            stream.release(stream)
+            raise RuntimeError("Error while accessing schema from stream.")
+
+        try:
+            create_fields_from_arrow_schema(destLayer, &schema, options, geometry_name)
+        except Exception as e:
+            schema.release(&schema)
+            stream.release(stream)
+            raise e
+
+        while True:
+            errcode = stream.get_next(stream, &array)
+            if errcode != 0:
+                schema.release(&schema)
+                stream.release(stream)
+                raise RuntimeError("Error while accessing batch from stream.")
+
+            # We've reached the end of the stream
+            if array.release == NULL:
+                break
+
+            errcode = OGR_L_WriteArrowBatch(destLayer, &schema, &array, options)
+            if not errcode:
+                if array.release != NULL:
+                    array.release(&array)
+
+                schema.release(&schema)
+                stream.release(stream)
+                raise RuntimeError("Error while writing batch to OGR layer.")
+
+            if array.release != NULL:
+                array.release(&array)
+
+        schema.release(&schema)
+        stream.release(stream)
+
+
+    cdef get_extension_metadata(const ArrowSchema* schema):
+        cdef const char *metadata = schema.metadata
+
+        extension_name = None
+        extension_metadata = None
+
+        if metadata == NULL:
+            return extension_name, extension_metadata
+
+        n = int.from_bytes(metadata[:4], byteorder=sys.byteorder)
+        pos = 4
+
+        for i in range(n):
+            length_key = int.from_bytes(metadata[pos:pos+4], byteorder=sys.byteorder)
+            pos += 4
+            key = metadata[pos:pos+length_key]
+            pos += length_key
+            length_value = int.from_bytes(metadata[pos:pos+4], byteorder=sys.byteorder)
+            pos += 4
+            value = metadata[pos:pos+length_value]
+            pos += length_value
+
+            if key == b"ARROW:extension:name":
+                extension_name = value
+            elif key == b"ARROW:extension:metadata":
+                extension_metadata = value
+
+        return extension_name, extension_metadata
+
+
+    cdef is_geometry_field(const ArrowSchema* schema):
+        name, _ = get_extension_metadata(schema)
+        if name is not None:
+            if name == b"geoarrow.wkb" or name == b"ogc.wkb":
+                return True
+            # raise an error for other geoarrow types
+            if name.startswith(b"geoarrow."):
+                raise NotImplementedError(
+                    f"Writing a geometry column of type {name.decode()} is not yet "
+                    "supported (only WKB is currently supported)."
+                )
+
+        return False
+
+
+    # Create output fields using CreateFieldFromArrowSchema()
+    cdef create_fields_from_arrow_schema(
+        OGRLayerH destLayer,
+        const ArrowSchema* schema,
+        char** options,
+        str geometry_name
+    ):
+        # The schema object is a struct type where each child is a column.
+        cdef ArrowSchema* child
+        for i in range(schema.n_children):
+            child = schema.children[i]
+            metadata = child.metadata
+
+            # Don't create property for geometry column
+            field_name = get_string(child.name)
+            if field_name == geometry_name:
+                continue
+
+            if is_geometry_field(child):
+                continue
+
+            errcode = OGR_L_CreateFieldFromArrowSchema(
+                destLayer, child, options)
+            if errcode == 0:
+                raise RuntimeError(
+                    f"error while creating field from Arrow for field {i} with name "
+                    f"{get_string(child.name)} and type {get_string(child.format)}"
+                )