+++ CHANGES.md
@@ -4,6 +4,9 @@
 
 ### Improvements
 
+-   Support for writing based on Arrow as the transfer mechanism of the data
+    from Python to GDAL (requires GDAL >= 3.8). This is provided through the
+    new `pyogrio.raw.write_arrow` function (#314, #346).
 -   `read_arrow` and `open_arrow` now provide
     [GeoArrow-compliant extension metadata](https://geoarrow.org/extension-types.html),
     including the CRS, when using GDAL 3.8 or higher (#366).